Removed interface PermissionTemplateManager
