+++ webcurator-core/src/test/java/org/webcurator/core/admin/MockPermissionTemplateManagerImpl.java
@@ -3,7 +3,7 @@ package org.webcurator.core.admin;
 import org.webcurator.auth.*;
 import org.webcurator.domain.*;
 
-public class MockPermissionTemplateManagerImpl extends PermissionTemplateManagerImpl {
+public class MockPermissionTemplateManagerImpl extends PermissionTemplateManager {
 
 	public MockPermissionTemplateManagerImpl(String filename) {
 		super();
